requirements.txt
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
DynamicHtml==1.0.3
favicon==0.7.0
filelock==3.16.1
Flask==3.0.3
googlesearch-python==1.2.5
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
latex2mathml==3.77.0
Markdown==3.7
MarkupSafe==2.1.5
mdtex2html==1.3.0
openai==1.51.0
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.1
pyppeteer==2.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
requests-file==2.1.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
websockets==10.4
Werkzeug==3.0.4
zipp==3.20.2
